pyproject.toml
[tool.poetry]
name = "dep"
version = "0.1.0"
description = ""
authors = ["legendarym4x <maxwel842@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
requests = "^2.31.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.6"
alembic = "^1.13.1"
jinja2 = "^3.1.3"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
redis = "^5.0.1"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"
pytest-mock = "^3.12.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
pytest = "7.4.4"
pytest-asyncio = "^0.23.4"
httpx = "^0.26.0"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"